+++ prefabfiles.lua
@@ -55,7 +55,6 @@ PrefabFiles = {
 	"refiner",
 	"scarecrow",	
 	"beanstalk_chunk",	
-	"magic_beans",
 	"beanstalk_sapling",	
 	"cloud_cotton",		
 	"cloud_turf",	
@@ -136,6 +135,8 @@ PrefabFiles = {
 	"beanlet_hut",
 	"campfirefire_dragon",
 	"thunder_pinecone",
+	"ua_cookedfoods",
+	"magic_beans",
 
 	--Stuff not started.
 	--------------------